Fix new acct by splitting entrypoint and CMD
Entrypoint is the executable, and CMD are the parameters passed into the
entrypoint. This means that we can run the private key generation as an
entrypoint, and then use `$@` in the start script to run the CMD as it's
own command / process. This is great because it allows creating new
accounts, mining, and validating to all use the same Docker image--just
different CMDs